feat: add limit & request for pods dashboard and disable throttle rule
The prometheus alert CPUThrottlingHigh is firing off all the time for several pods. Seems to be a storage or two and metrics exporters.

Read more about the issue here. Seems likle Kubernetes throttles some pods without the need to. I've confirmed this in the dashboard. Took a sample from the list of throttled services and verified that the pod were way below request and limit..

https://devops.stackexchange.com/questions/6494/prometheus-alert-cputhrottlinghigh-raised-but-monitoring-does-not-show-it/6753#6753

Have to either wait for a kernel patch or disable CFS quotas which I'm not sure how we can do in GKE.

Adding severity none on the alert for now.